requirements.txt
# Web Framework
Flask==2.3.3
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
Werkzeug==2.3.7
gunicorn==21.2.0

# Data Processing
pandas==2.1.0
numpy==1.25.2
akshare==1.16.39
tqdm==4.66.1

# Finance & Analysis
# ta-lib==0.4.26  # 可能需要单独安装，取决于系统
matplotlib==3.7.3
scikit-learn==1.3.0

# Configuration
PyYAML==6.0.1
python-dotenv==1.0.0

# Database (Optional)
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7  # PostgreSQL
mysqlclient==2.2.0  # MySQL
aiosqlite==0.19.0  # SQLite

# Caching
cachetools==5.3.1

# Utilities
requests==2.31.0
pytz==2023.3
tenacity==8.2.3  # 重试机制
apscheduler==3.10.4  # 定时任务

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Monitoring
prometheus-client==0.17.1

# Security
PyJWT==2.8.0
passlib==1.7.4

# Development Tools
black==23.7.0
flake8==6.1.0
mypy==1.5.1